+++ application/core/LSYii_Controller.php
@@ -121,11 +121,15 @@ abstract class LSYii_Controller extends CController
 
         if (Yii::app()->getConfig('debug')==1) {//For debug purposes - switch on in config.php
             @ini_set("display_errors", 1);
-            error_reporting(E_ALL);
+            error_reporting(E_ERROR | E_WARNING | E_PARSE);
         }
         elseif (Yii::app()->getConfig('debug')==2) {//For debug purposes - switch on in config.php
             @ini_set("display_errors", 1);
             error_reporting(E_ALL | E_STRICT);
+        }
+        elseif(Yii::app()->getConfig('debug')>0)
+        {
+            @ini_set("display_errors", 0);
         }
         else {
             @ini_set("display_errors", 0);