+++ application/helpers/expressions/em_manager_helper.php
@@ -720,6 +720,9 @@
             $_SESSION['LEMforceRefresh'] = true;// For Expression manager string
             /* Bug #09589 : update a survey don't reset actual test => Force reloading of survey */
             $iSessionSurveyId=self::getLEMsurveyId();
+            /* Remove state of DB error too */
+            Yii::app()->user->setState('dberror_'.$iSessionSurveyId,null);
+
             if($aSessionSurvey=Yii::app()->session["survey_{$iSessionSurveyId}"])
             {
                 $aSessionSurvey['LEMtokenResume']=true;
@@ -4803,6 +4806,8 @@
         */
         static function StartSurvey($surveyid,$surveyMode='group',$aSurveyOptions=NULL,$forceRefresh=false,$debugLevel=0)
         {
+            /* If we start a survey : we need to remove DB error */
+            Yii::app()->user->setState('dberror_'.$surveyid,null);
             $LEM =& LimeExpressionManager::singleton();
             $LEM->sid=sanitize_int($surveyid);
             $LEM->sessid = 'survey_' . $LEM->sid;
@@ -5400,12 +5405,18 @@
                             break;
                         case 'N': //NUMERICAL QUESTION TYPE
                         case 'K': //MULTIPLE NUMERICAL QUESTION
-                            if (trim($val)=='')
+                            if (trim($val)=='' || !is_numeric($val)) // is_numeric error is done by EM : then show an error and same page again
                             {
                                 $val=NULL;  // since some databases can't store blanks in numerical inputs
                             }
+                            elseif(!preg_match("/^[-]?(\d{1,20}\.\d{0,10}|\d{1,20})$/",$val)) // DECIMAL(30,10)
+                            {
+                                // Here : we must ADD a message for the user and set the question "not valid" : show the same page + show with input-error class
+                                $val=NULL;
+                            }
                             break;
                         default:
+                            // @todo : control length of DB string, if answers in single choice is valid too (for example) ?
                             break;
                     }
 
@@ -5428,8 +5439,8 @@
 
                     if (!dbExecuteAssoc($query))
                     {
-                        echo submitfailed('');  // TODO - report SQL error?
-
+                        $flashMessage= submitfailed('', $query);  // TODO - report SQL error?
+                        Yii::app()->user->setState('dberror_'.$this->sid, $flashMessage);
                         if (($this->debugLevel & LEM_DEBUG_VALIDATION_SUMMARY) == LEM_DEBUG_VALIDATION_SUMMARY) {
                             $message .= $this->gT('Error in SQL update');  // TODO - add  SQL error?
                         }